Moving transform to be typescript
